pyproject.toml
[project]
name = "elevenlabs-mcp-server"
version = "0.1.0"
description = "MCP server for ElevenLabs text-to-speech integration"
requires-python = ">=3.11"
dependencies = [
    "mcp",
    "requests",
    "pydub",
    "python-dotenv",
    "pytest>=8.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elevenlabs_mcp"]

[project.scripts]
elevenlabs-mcp = "elevenlabs_mcp.server:main"
